requirements.txt
Django>=3.2,<3.3
sorl-thumbnail>=12.0,<13.0